chore(get-prop-by-agent):refactor property models
