+++ api/models/users.js
@@ -35,7 +35,18 @@ class User {
       : `SELECT * FROM property WHERE status='available'`
     );
   }
+  
+  static getPropertyByOwner(email){
+    return client.query(`SELECT * FROM property WHERE owner='${email}'`)
+  }
 
+  static checkIfPropertyExist(property){
+    return client.query(`SELECT * FROM property WHERE ${property}`)
+  }
+
+  static queryAll(query){
+      return client.query(`SELECT * FROM property WHERE ${query}`)
+  }
 }
 
 class Agent extends User {
